# public/openapi-everything.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AB Brain - Complete API",
...
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "https://ab-brain.up.railway.app/api"
    }
  ],
  "paths": {
...
    "/workouts/{id}": {
      "get": {
        "operationId": "getWorkout",
        "summary": "Get a single workout by ID",
        "responses": {
          "200": {
            "description": "Single workout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateWorkout",
        "summary": "Update a workout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Workout"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated workout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteWorkout",
        "summary": "Delete a workout",
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
...
    "/workouts": {
      "get": {
        "operationId": "listWorkouts",
        "summary": "List or search workout logs",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Full-text search query"
          },
          {
            "name": "workout_type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "hill",
                "strength",
                "run",
                "hybrid",
                "recovery",
                "ruck"
              ]
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by tag (e.g. spartan, carry)"
          },
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Workouts on or after this date"
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Workouts before this date"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 30
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workout log entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "workouts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Workout"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "logWorkout",
        "summary": "Log a new Spartan workout with structured fields for type, workout details, metrics, performance, body feedback, and adjustments",
        "description": "Post a structured workout log. ChatGPT should use this after a workout recap conversation. Include workout_type, focus, warmup, main_sets, carries, metrics (time, distance, elevation), effort rating 1-10, performance notes (slowdown, failure point), body feedback (grip, legs, cardio, shoulder), a...",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "workout_type"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Auto-generated if empty: 'Spartan Workout \u2013 DATE \u2013 TYPE'"
                  },
                  "workout_date": {
                    "type": "string",
                    "format": "date",
                    "description": "YYYY-MM-DD, defaults to today"
                  },
                  "workout_type": {
                    "type": "string",
                    "enum": [
                      "hill",
                      "strength",
                      "run",
                      "hybrid",
                      "recovery",
                      "ruck"
                    ]
                  },
                  "location": {
                    "type": "string",
                    "description": "e.g. Runyon Canyon, Home Gym"
                  },
                  "elevation": {
                    "type": "string",
                    "description": "Base elevation if relevant, e.g. 1,320 ft"
                  },
                  "focus": {
                    "type": "string",
                    "description": "Goal of the workout, e.g. hill durability, grip endurance"
                  },
                  "warmup": {
                    "type": "string",
                    "description": "Warm-up activities"
                  },
                  "main_sets": {
                    "type": "string",
                    "description": "Main workout sets and exercises"
                  },
                  "carries": {
                    "type": "string",
                    "description": "Carries, lifts, or loaded movements"
                  },
                  "time_duration": {
                    "type": "string",
                    "description": "Text duration (e.g. '45 min', '1:15:00'). Server auto-parses to duration_minutes."
                  },
                  "duration_minutes": {
                    "type": "integer",
                    "description": "Duration in minutes. Auto-computed from time_duration if not provided."
                  },
                  "distance": {
                    "type": "string",
                    "description": "Text distance (e.g. '3.2 mi'). Server auto-parses to distance_value."
                  },
                  "distance_value": {
                    "type": "number",
                    "description": "Distance as number. Auto-computed from distance if not provided."
                  },
                  "elevation_gain": {
                    "type": "string",
                    "description": "Text elevation (e.g. '850 ft'). Server auto-parses to elevation_gain_ft."
                  },
                  "elevation_gain_ft": {
                    "type": "integer",
                    "description": "Elevation gain in feet. Auto-computed from elevation_gain if not provided."
                  },
                  "effort": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10,
                    "description": "Effort rating 1-10"
                  },
                  "slowdown_notes": {
                    "type": "string",
                    "description": "Where did I slow down or break form"
                  },
                  "failure_first": {
                    "type": "string",
                    "description": "What failed first: legs / grip / cardio / shoulders"
                  },
                  "grip_feedback": {
                    "type": "string",
                    "description": "Grip status: solid / pumped / gave out"
                  },
                  "legs_feedback": {
                    "type": "string",
                    "description": "Legs status: strong / heavy / cramped"
                  },
                  "cardio_feedback": {
                    "type": "string",
                    "description": "Cardio status: steady / gasping / recovered fast"
                  },
                  "shoulder_feedback": {
                    "type": "string",
                    "description": "Shoulder status: fine / tight / clicking"
                  },
                  "body_notes": {
                    "type": "string",
                    "description": "Additional body feedback or recovery notes"
                  },
                  "adjustment": {
                    "type": "string",
                    "description": "What to change next time: weight, pace, structure"
                  },
                  "heart_rate_avg": {
                    "type": "integer",
                    "description": "Average heart rate in bpm"
                  },
                  "heart_rate_max": {
                    "type": "integer",
                    "description": "Max heart rate in bpm"
                  },
                  "pace_avg": {
                    "type": "string",
                    "description": "Average pace, e.g. 8:30/mi"
                  },
                  "cadence_avg": {
                    "type": "number",
                    "description": "Average cadence (steps per minute)"
                  },
                  "active_calories": {
                    "type": "number",
                    "description": "Active calories burned"
                  },
                  "total_calories": {
                    "type": "number",
                    "description": "Total calories burned"
                  },
                  "splits": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    },
                    "description": "Split/lap data"
                  },
                  "exercises": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    },
                    "description": "Individual exercises with sets/reps/weight"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Tags without # prefix, e.g. [\"spartan\", \"hill\", \"carry\"]"
                  },
                  "ai_source": {
                    "type": "string",
                    "default": "chatgpt"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout logged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Workout": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "workout_date": {
            "type": "string",
            "format": "date"
          },
          "workout_type": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "elevation": {
            "type": "string"
          },
          "focus": {
            "type": "string"
          },
          "warmup": {
            "type": "string"
          },
          "main_sets": {
            "type": "string"
          },
          "carries": {
            "type": "string"
          },
          "time_duration": {
            "type": "string"
          },
          "duration_minutes": {
            "type": "integer"
          },
          "distance": {
            "type": "string"
          },
          "distance_value": {
            "type": "number"
          },
          "elevation_gain": {
            "type": "string"
          },
          "elevation_gain_ft": {
            "type": "integer"
          },
          "effort": {
            "type": "integer"
          },
          "slowdown_notes": {
            "type": "string"
          },
          "failure_first": {
            "type": "string"
          },
          "grip_feedback": {
            "type": "string"
          },
          "legs_feedback": {
            "type": "string"
          },
          "cardio_feedback": {
            "type": "string"
          },
          "shoulder_feedback": {
            "type": "string"
          },
          "body_notes": {
            "type": "string"
          },
          "adjustment": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "heart_rate_avg": {
            "type": "integer"
          },
          "hr_avg": {
            "type": "integer"
          },
          "heart_rate_max": {
            "type": "integer"
          },
          "hr_max": {
            "type": "integer"
          },
          "pace_avg": {
            "type": "string"
          },
          "cadence_avg": {
            "type": "number"
          },
          "cadence": {
            "type": "integer"
          },
          "active_calories": {
            "type": "number"
          },
          "cal_active": {
            "type": "integer"
          },
          "total_calories": {
            "type": "number"
          },
          "cal_total": {
            "type": "integer"
          },
          "splits": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "exercises": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}